+++ packages/spotlight/src/ui/telemetry/components/metrics/MetricsList.tsx
@@ -24,6 +24,7 @@ import { getFormattedNumber } from "../../utils/duration";
 import { aggregateMetrics, calculatePercentiles, groupMetricsByName } from "../../utils/metrics";
 import { truncateId } from "../../utils/text";
 import MetricDetail from "./MetricDetail";
+import MetricTypeBadge from "./components/MetricTypeBadge";
 
 type MetricsListProps = {
   traceId?: string;
@@ -368,25 +369,6 @@ function MetricSampleItem({
         {sample.span_id && <span className="text-primary-500 ml-3 font-mono">span {truncateId(sample.span_id)}</span>}
       </div>
     </Link>
-  );
-}
-
-function MetricTypeBadge({ type }: { type: MetricType }) {
-  const colors: Record<MetricType, string> = {
-    counter: "bg-blue-600/30 text-blue-300 border-blue-500/30",
-    gauge: "bg-green-600/30 text-green-300 border-green-500/30",
-    distribution: "bg-purple-600/30 text-purple-300 border-purple-500/30",
-  };
-
-  return (
-    <span
-      className={cn(
-        "rounded border px-1.5 py-0.5 text-xs uppercase",
-        colors[type] || "bg-primary-700 text-primary-100",
-      )}
-    >
-      {type}
-    </span>
   );
 }
